commmon: fastboot: cast 64bit when converting sectors to bytes
This patch fixes a cast error when converting sectors to bytes. The
start and size types of the disk_partition_t are 32bit until the
SYS_64BIT_LBA is defined. To prevent the converting error, we should
cast the value to 64bit values.

+++ common/cmd_fastboot.c
@@ -1774,8 +1774,10 @@ static int get_mmc_partition_tables()
 			break;
 
 		strcpy(ptable[pcount].name, info.name);
-		ptable[pcount].start = info.start * info.blksz;
-		ptable[pcount].length = info.size * info.blksz;
+		ptable[pcount].start = (unsigned long long)info.start *
+							   info.blksz;
+		ptable[pcount].length = (unsigned long long)info.size *
+							   info.blksz;
 		ptable[pcount].flags = FASTBOOT_PTENTRY_FLAGS_USE_MMC_CMD;
 #ifdef CONFIG_FASTBOOT_FLASH_CHUNK
 		ptable[pcount].flags |= FASTBOOT_PTENTRY_FLAGS_FLASH_CHUNK;
